Search Feature Added

+++ src/main/java/com/coviam/blabla/merchant/dao/MerchantRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import com.coviam.blabla.merchant.entity.Merchant;
 
 @Repository
-public interface MerchantRepository extends CrudRepository<Merchant,Long> {
+public interface MerchantRepository extends CrudRepository<Merchant,Integer> {
 }
