+++ src/main/java/com/coviam/blabla/merchant/entity/ScoreId.java
@@ -8,22 +8,22 @@ import javax.persistence.Embeddable;
 @Embeddable
 public class ScoreId implements Serializable{
 	@Column(name="merchant_id")
-	long merchantId;
+	int merchantId;
 	@Column(name="product_id")
-	long productId;
+	int productId;
 	public ScoreId() {
 		super();
 		}
-	public long getMerchantId() {
+	public int getMerchantId() {
 		return merchantId;
 	}
-	public void setMerchantId(long merchantId) {
+	public void setMerchantId(int merchantId) {
 		this.merchantId = merchantId;
 	}
-	public long getProductId() {
+	public int getProductId() {
 		return productId;
 	}
-	public void setProductId(long productId) {
+	public void setProductId(int productId) {
 		this.productId = productId;
 	}
 }
